+++ docker-compose.dev.yaml
@@ -8,7 +8,7 @@ services:
       # args:
       #   NODE_ENV: development
     volumes:
-      - ./:/app:ro
+      # - ./:/app:ro
       - /app/node_modules
     env_file:
       - ./.env
