+++ src/config/allowedOrigins.config.ts
@@ -1,5 +1,6 @@
 const allowedOrigins = [
-    "http://localhost:3000",
-]
+  "http://localhost:3000/*",
+  "https://mydoshbox.vercel.app/*",
+];
 
 export default allowedOrigins;
