+++ src/modules/authentication/organizationUserAuth/organizationUser/organizationAuth.controller.ts
@@ -5,7 +5,7 @@ import crypto from "crypto";
 import catchAsync from "../../../../utilities/catchAsync";
 import AppError from "../../../../utilities/appError";
 import { sendVerificationEmail } from "../../../../utilities/email.utils";
-import IndividualUser from "../../individualUserAuth/individualUserAuth.model";
+import IndividualUser from "../../individualUserAuth/individualUserAuth.model1";
 
 interface TokenPayload {
   id: string;
@@ -33,11 +33,7 @@ const createSendToken = (user: any, statusCode: number, res: Response) => {
   });
 };
 
-export const organizationUserSignup = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
+export const organizationUserSignup = async (req: Request, res: Response) => {
   try {
     const {
       organization_name,
@@ -83,7 +79,7 @@ export const organizationUserSignup = async (
     if (password !== password_confirmation) {
       return res.status(401).json({
         status: "fail",
-        message: "Password do not match",
+        message: "Passwords do not match",
       });
     }
 
@@ -101,7 +97,8 @@ export const organizationUserSignup = async (
       });
     }
 
-    const org = await OrganizationModel.create({
+    // Create a new user
+    const newUser = new OrganizationModel({
       organization_name,
       organization_email,
       contact_email,
@@ -110,9 +107,12 @@ export const organizationUserSignup = async (
       role: "org",
     });
 
+    // Save the user to the database
+    await newUser.save();
+
     const verificationToken = jwt.sign(
       {
-        email: org.organization_email,
+        email: newUser.organization_email,
       },
       process.env.JWT_SECRET as string,
       {
@@ -120,7 +120,7 @@ export const organizationUserSignup = async (
       }
     );
 
-    await sendVerificationEmail(org.organization_email, verificationToken);
+    await sendVerificationEmail(organization_email, verificationToken);
 
     return res.status(201).json({
       status: "true",
@@ -128,116 +128,13 @@ export const organizationUserSignup = async (
         "Account successfully created. Verification email sent. Verify account to continue",
     });
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  } catch (err: any) {
-    next(err);
+  } catch (error: any) {
+    console.error("Error registering the user:", error);
+    return res
+      .status(500)
+      .json({ message: "Error registering the user", error });
   }
 };
-
-// export const organizationUserLogin = async (
-//   req: Request,
-//   res: Response,
-//   next: NextFunction
-// ) => {
-//   try {
-//     const { organization_email, password } = req.body;
-
-//     if (!organization_email || !password) {
-//       res.status(401).json({
-//         status: "fail",
-//         message: "Password and organization email are required",
-//       });
-//     }
-
-//     // 2) Check if user exists && password is correct
-//     const loggedInUser = await OrganizationModel.findOne({
-//       organization_email,
-//     }).select("+password");
-
-//     if (
-//       !loggedInUser ||
-//       !(await loggedInUser.correctPassword(password, loggedInUser.password))
-//     ) {
-//       return res.status(401).json({
-//         status: "fail",
-//         message: "Incorrect details",
-//       });
-//     }
-
-//     if (!loggedInUser.email_verified) {
-//       const verificationToken = jwt.sign(
-//         {
-//           email: loggedInUser.organization_email,
-//         },
-//         process.env.JWT_SECRET as string,
-//         {
-//           expiresIn: 60 * 60,
-//         }
-//       );
-
-//       await sendVerificationEmail(
-//         loggedInUser.organization_email,
-//         verificationToken
-//       );
-
-//       // Send a response
-//       return res.status(200).json({
-//         status: "true",
-//         message:
-//           "Account is unverified! Verification email sent. Verify account to continue",
-//       });
-//     }
-
-//     // 3) If everything ok, send token to client
-//     const createSessionAndSendTokensOptions = {
-//       user: loggedInUser.toObject(),
-//       userAgent: req.get("user-agent") || "",
-//       role: loggedInUser.role,
-//       message: "Organization user sucessfully logged in",
-//     };
-
-//     const { status, message, user, accessToken, refreshToken } =
-//       await createSessionAndSendTokens(createSessionAndSendTokensOptions);
-
-//     return res.status(200).json({
-//       status,
-//       message,
-//       user,
-//       refreshToken,
-//       accessToken,
-//     });
-//     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-//   } catch (err: any) {
-//     next(err);
-//   }
-// };
-
-// export const OrganizationUserForgotPassword = catchAsync(
-//   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-//     // 1) Get user based on POSTed email
-//     const org = await OrganizationModel.findOne({
-//       email: req.body.email,
-//     });
-//     if (!org) {
-//       return next(new AppError("There is no user with email address.", 404));
-//     }
-
-//     // 2) Generate the random reset token
-//     const resetToken = org.createPasswordResetToken();
-//     await org.save({ validateBeforeSave: false });
-
-//     // 3) Send it to user's email
-//     const resetURL = `${req.protocol}://${req.get(
-//       "host"
-//     )}/auth/organization/resetPassword/${resetToken}`;
-
-//     try {
-//       sendURLEmail([org.organization_email], resetURL);
-//       res.status(200).json({ message: "success" });
-//     } catch (err) {
-//       return next(new AppError("There is an error sending the email.", 500));
-//     }
-//   }
-// );
 
 export const organizationUserResetPassword = catchAsync(
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
