+++ src/modules/sessions/session.model.ts
@@ -1,8 +1,4 @@
 import mongoose, { Schema } from "mongoose";
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-// import GoogleOrganizationUser from "../authentication/organizationUserAuth/googleOrganizationUserAuth.model";
-// // eslint-disable-next-line @typescript-eslint/no-unused-vars
-// import GoogleIndividualUser from "../authentication/individualUserAuth/googleIndividualAuth.model";
 
 interface ISession {
   user: Schema.Types.ObjectId;
