+++ src/modules/users/individualUsers/individualUsers.route.ts
@@ -1,14 +1,13 @@
 import { Router } from "express";
-import {
-  getIndividualUser,
-  getAllIndividualUsers,
-  updateIndividualUser,
-  deleteIndividualUser,
-} from "./individualUsers.controller";
+import // getIndividualUser,
+// getAllIndividualUsers,
+// updateIndividualUser,
+// deleteIndividualUser,
+"./individualUsers.controller";
 
 const router = Router();
 
-router.get("/my-profile", getIndividualUser);
-router.patch("/update-my-profile", updateIndividualUser);
+// router.get("/my-profile", getIndividualUser);
+// router.patch("/update-my-profile", updateIndividualUser);
 
 export default router;
