+++ src/utilities/validator.utils.ts
@@ -1,8 +1,8 @@
 interface Validator {
-  (value: any): boolean;
+  (value: string): boolean;
 }
 
-export const emailValidator: Validator = (value: any): boolean => {
+export const emailValidator: Validator = (value: string): boolean => {
   // Regular expression for validating email addresses
   const emailRegex: RegExp = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
   return emailRegex.test(value);
